% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_condition_tag}
\alias{add_condition_tag}
\title{Add Condition Tag To Data Frame}
\usage{
add_condition_tag(df, parameters)
}
\arguments{
\item{df}{Data Frame or Tibble, with the following columns:
\itemize{
\item sweep
}}

\item{parameters}{Data Frame or Tibble with experiment parameters, expected columns are as follows:
\itemize{
\item condition_start - Integer
\item condition_end - Integer
\item condition - Character
}}
}
\value{
Data Frame or Tibble with the additional column: \code{condition}.
}
\description{
Adds a condition tag to the data frame allowing analysis based upon the tag
}
\details{
Uncategorized events will default to \code{NA}.
}
\examples{
simple_tbl <- tibble::tibble(sweep = as.ordered(rep(1:5, each = 2)))
user_drug_args <-
  tibble::tribble(
    ~condition_start, ~condition_end, ~condition,
    1, 2, "control",
    4, 5, "drug_1"
  )

add_condition_tag(
  df = simple_tbl,
  parameters = user_drug_args
)

}
\seealso{
\itemize{
\item \link{add_sweep_number_to_rows}
\item \link{insert_rows_for_missing_sweeps}
}
}
