% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_event_index}
\alias{add_event_index}
\title{Add Event Index To Data Frame}
\usage{
add_event_index(df)
}
\arguments{
\item{df}{Data Frame or Tibble with the following columns:
\itemize{
\item amplitude
\item sweep
\item stimulus
}}
}
\value{
Data Frame or Tibble with event index numbers in column the \code{event_index}.
}
\description{
Several events may occur in response to a experimental stimulus, this function will add a sub-index
allowing for segregation by the i-th event to a n-th stimulus.
}
\details{
The first event for a given stimulus will always start with an index value of 1. If multiple events occur,
and iterative index is created. If no events occur, the index takes a \code{NA} value.
}
\examples{
simple_tbl <- tibble::tibble(
  sweep = as.ordered(c(1, 1, 1, 1, 2, 2, 2, 2, 2)),
  time_ms = c(405, 408, NA, 445, seq(405, 445, 10)),
  amplitude = c(100, 100, NA, 100, 100, 100, 100, 100, 100),
  stimulus = ordered(c(1, 1, 2, "r", 1, 2, "o", "o", "r"), levels = c(1, 2, "o", "r"))
)
add_event_index(simple_tbl)

}
\seealso{
\itemize{
\item \link{add_stimulus_index}
}
}
