% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_event_jitter}
\alias{add_event_jitter}
\title{Calculate Event Jitter From Stimulus}
\usage{
add_event_jitter(df, to_rise = FALSE)
}
\arguments{
\item{df}{Data Frame or Tibble, with the following columns:
\itemize{
\item time_ms
\item stim_times
}}

\item{to_rise}{Logical, default is \code{FALSE} and jitter is calculated between the stimulus time to the peak
amplitude, i.e. \code{time_ms}. If \code{TRUE} the rise time, \code{rise_ms}, will be subtracted from the time
of peak amplitude, thus calculating jitter to the start the 20/80 rise time.}
}
\value{
Data Frame or Tibble with numeric column, \code{jitter}, in milliseconds.
}
\description{
The jitter of an synaptic event (i.e. action potential, post-synaptic event) is calculated using
difference between the stimulus time and the peak amplitude time (or start of the 20/80 rise).
}
\examples{
simple_df <- data.frame(
  sweep = as.ordered(rep(1:2, c(4, 5))),
  time_ms = c(405, 408, NA, 445, seq(405, 445, 10)),
  rise_ms = rep(0.5, 9),
  stimulus = ordered(c(1, 1, 2, "r", 1, 2, "o", "o", "r"), levels = c(1, 2, "o", "r")),
  stim_times = c(400, 400, 410, 440, 400, 410, NA, NA, 440)
)

add_event_jitter(simple_df)

}
