% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_normalized_amplitude}
\alias{add_normalized_amplitude}
\title{Calculate The Normalized Amplitude}
\usage{
add_normalized_amplitude(
  df,
  normalize_condition = "control",
  normalize_stimulus = 1
)
}
\arguments{
\item{df}{Data Frame or Tibble, with the following columns:
\itemize{
\item sweep
\item amplitude
\item condition
\item stimulus
\item event_index
}}

\item{normalize_condition}{Character, the condition to normalize all amplitudes against, defaults to 'control'.}

\item{normalize_stimulus}{Numeric, the stimulus to normalize against, defaults to '1'.}
}
\value{
Data Frame or Tibble with the additional column: \code{amplitude_normalized}.
}
\description{
This will calculate a normalized amplitude based upon a user defined baseline - set using the args:
\code{normalize_condition} and \code{normalize_stimulus}. The baseline is calculated from the mean amplitude of
first "event" after the user-defined \code{normalize_stimulus} value.
}
\examples{
simple_df <-
  data.frame(
    sweep = c(1:20),
    amplitude = rep(c(100, 50), each = 10),
    condition = rep(c("control", "drug_1"), c(10, 10)),
    stimulus = 1,
    event_index = 1
  )

add_normalized_amplitude(simple_df,
  normalize_condition = "control",
  normalize_stimulus = 1
)

}
