% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_ppr}
\alias{add_ppr}
\title{Add Pair-Pulse Ratio}
\usage{
add_ppr(df, column_ref = "amplitude")
}
\arguments{
\item{df}{Data Frame or Tibble with the following columns:
\itemize{
\item sweep
\item condition
\item stimulus
\item event_index
\item amplitude OR another reference column
}}

\item{column_ref}{Character, column to use for ppr calculation. Defaults to \code{"amplitude"}.}
}
\value{
Data Frame or Tibble with the additional column: \code{condition}.
}
\description{
This will add the pair-pulse ratio value to rows where the stimulus == 2 and event_index == 1. For paired stimuli
with no event on the 2nd stimulus, \code{0} is returned; for stimuli with no event on the 1st stimulus, \code{NA} is
returned.
}
\examples{
n <- 2
simple_df <-
  data.frame(
    sweep = rep(1:(4 * n), each = 2),
    condition = rep(c(NA, "control", "drug_1", "drug_2"), each = (2 * n)),
    amplitude = c(rep(c(40, 20), n), rep(c(40, 20), n), rep(c(10, NA), n), rep(c(NA, 30), n)),
    stimulus = rep(1:2, 4 * n),
    event_index = 1
  )

add_ppr(df = simple_df, column_ref = "amplitude")

}
