% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_stimulus_index}
\alias{add_stimulus_index}
\title{Add Stimulus Index To Data Frame}
\usage{
add_stimulus_index(
  df,
  time_of_stim,
  isi,
  sweep_count,
  recovery_stim = NULL,
  add_missing = TRUE
)
}
\arguments{
\item{df}{Data Frame or Tibble with the following columns:
\itemize{
\item sweep
\item time_ms
}}

\item{time_of_stim}{List, with event times in milliseconds.}

\item{isi}{Numeric, interstimulus interval in milliseconds.}

\item{sweep_count}{Numeric, total number of experiment sweeps/trials.}

\item{recovery_stim}{Numeric, default is \code{NULL}. Time of the recovery stimulus in milliseconds, useful in
paired-pulse ratio experiments when a recovery stimulus is given to measure the rate of recovery.}

\item{add_missing}{Logical, default is \code{TRUE}. Rows for missing events will be inserted with corresponding
sweep and stimulus number.}
}
\value{
Data Frame or Tibble with the following columns:
\itemize{
\item \code{stimulus}, Factor column with the stimulus index.
\item \code{stim_time}, Numeric column with stimulus times.
}
}
\description{
Categorizes events by relationship to a given stimulus, creating a stimulus index.
}
\details{
Determines which events fall within a stimulus window, defined by the interstimulus interval (isi), and
adds rows with \code{NA} for missing events. The \code{stimulus} column can take on several values:
\itemize{
\item Character int, e.g. \code{1}, representing an event occurring within a stimulus window.
\item \code{NA}, representing events that do not correspond to a stimulus.
\item Character, \code{r}, representing the recovery stimulus. If \code{recovery_stim} is not \code{NULL}.
}
}
\note{
Run \link{insert_rows_for_missing_sweeps} before this function, as it will only insert NA for missing events on
sweep numbers already present the data frame. A check occurs in this function and notifies the user.
}
\examples{
simple_df <- data.frame(
  sweep = as.ordered(rep(1:2, c(1, 2))),
  time_ms = c(405, 405, 415)
)

# add rows for missing events
add_stimulus_index(
  df = simple_df,
  time_of_stim = c(400, 410),
  isi = 10,
  sweep_count = 2,
  recovery_stim = NULL,
  add_missing = TRUE
)

}
\seealso{
\itemize{
\item \link{insert_rows_for_missing_sweeps}
\item \link{add_event_index}
}
}
