% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_sweep_number_to_rows}
\alias{add_sweep_number_to_rows}
\title{Add Sweep Number To Rows}
\usage{
add_sweep_number_to_rows(
  df,
  sweep_duration,
  sweep_count,
  time_ref = "rec_time_ms"
)
}
\arguments{
\item{df}{Data Frame or Tibble}

\item{sweep_duration}{Numeric, sweep/trial duration in seconds.}

\item{sweep_count}{Numeric, total number of experiment sweeps/trials.}

\item{time_ref}{Character, event time reference point column name, i.e. \code{rec_time_ms}.}
}
\value{
Data Frame or Tibble with the following columns:
\itemize{
\item \code{sweep}, Ordered Factor
}
}
\description{
Determines the sweep/trial number based upon the event time, \code{rec_time_ms}.
}
\examples{
simple_df <- data.frame(rec_time_ms = c("400.50", "4,400.50"))

add_sweep_number_to_rows(
  df = simple_df,
  sweep_duration = 1,
  sweep_count = 5,
  time_ref = "rec_time_ms"
)

\dontrun{
df <- add_sweep_number_to_rows(
  df = data,
  sweep_duration = matools_env$sweep_duration_sec,
  sweep_count = matools_env$sweeps_total,
  time_ref = "rec_time_ms"
)
}

}
\seealso{
\itemize{
\item \link{insert_rows_for_missing_sweeps}
}
}
