% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{align_condition_sweeps}
\alias{align_condition_sweeps}
\title{Align Condition Sweeps}
\usage{
align_condition_sweeps(df, condition_names)
}
\arguments{
\item{df}{Data Frame or Tibble, with the following columns:
* sweep
* condition}

\item{condition_names}{Character vector with condition names.}
}
\value{
Data Frame or Tibble, with the additional column:
* trial
}
\description{
Creates a 'trial' column and aligns the first sweep in each condition to trial '1'. The original
sweep value is not overwritten.
}
\seealso{
* \link{sequential_condition_sweeps}
* \link{plot_heatmap}
}
