% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{alignments2gw}
\alias{alignments2gw}
\title{alignments2gw}
\usage{
alignments2gw(alignments, drop = 0, ignore.overlaps = TRUE, verbose = TRUE)
}
\arguments{
\item{alignments}{GRanges or GRangesList of pooled reads}

\item{drop}{(default = 0) integer for dropping nodes in read space}

\item{ignore.overlaps}{(default = T) logical for whether to ignore overlaps in alignments when lifting to reference}

\item{verbose}{(default = T)}
}
\value{
gWalk object with nodes and edges lifted to reference coordinates
}
\description{
taken from alignments2gg, spits out intermediate gWalk object w/ nodes + edges lifted to ref coords
}
\author{
Marcin Imielinski, Joe DeRose, Xiaotong Yao, andrew ma
}
