% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{analysis_evoked_psp}
\alias{analysis_evoked_psp}
\title{Workflow: Analyze Evoked Post-Synaptic Potential Data}
\usage{
analysis_evoked_psp(data_folder_path, data_parameters)
}
\arguments{
\item{data_folder_path}{path of data folder containing *.ASC files}

\item{data_parameters}{path of csv file with project parameters}
}
\value{
tibble of aggregated data
}
\description{
Working example of how to use the functions within this package to create an analysis pipeline.
}
\details{
This particular workflow is used to analyze the amplitude, failure rate, potency, paired-pulse ratio of
post-synaptic potentials recorded using a intracellular electrophysiology technique.
}
\note{
This workflow has hard-coded values, yet is intended to serve as a reference point.
}
\examples{
\dontrun{
data_path <- paste(getwd(), "data-raw/evoked_psp", sep = "/")
data_folder <- paste(data_path, "data", sep = "/")
data_parameters <- paste(data_path, "evoked_psp_parameters.csv", sep = "/")

data_collection <- analysis_evoked_psp(
  data_folder_path = data_folder,
  data_parameters = data_parameters
)
}

}
