% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asc_lookup_table}
\alias{asc_lookup_table}
\title{ASC File LUT}
\usage{
asc_lookup_table()
}
\value{
tibble with three columns:
* name
* df_class - classes for data frames
* tbl_class - classes for tibbles
}
\description{
A lookup table with the column names and classes for ASC files generated by mini analysis
}
\note{
Several data frame columns are set to the base type "character" to avoid generating \code{NA} on number values
containing commas and periods, e.g. \code{1,200.00}, on import.
}
