% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asc_to_df}
\alias{asc_to_df}
\title{Convert .ASC File Into Data Frame}
\usage{
asc_to_df(file_path, merge_iei = FALSE)
}
\arguments{
\item{file_path}{Character, path to file. User can call \code{file.choose} to interactively select the file.}

\item{merge_iei}{Logical, if TRUE .asc and .txt files with matching names will be merged. Default is \code{FALSE}.}
}
\value{
data.frame, column names match default columns within Synaptosoft's Mini Analysis software.
Column names are as follows:
\itemize{
\item \code{ma_row}: Integer, corresponds to the row number generated in Mini Analysis. It is not recommended for use
as a reference/key value.
\item \code{rec_time_ms}: Character, continuous time value that is determined from the recording digitization value.
\item \code{amplitude} Double, amplitude of the detected event.
\item \code{rise_ms} Double, rise time in milliseconds.
\item \code{decay_ms} Double, decay time in milliseconds.
\item \code{area} Character
\item \code{baseline} Double
\item \code{noise} Double
\item \code{group}: Integer (0 to 255), user-assigned group number within MA.
\item \code{channel}: Integer, corresponds to the recording channel as defined by the user's DAC and digitizer.
\item \code{x10_90_rise} Double
\item \code{halfwidth} Double
\item \code{rise50} Double
\item \code{peak_direction} Integer
\item \code{burst} Integer
\item \code{burste} Integer
\item \code{x10_90slope} Character
\item \code{rel_time} Double
\item \code{iei}: Double, inter-event interval in milliseconds calculated in MA. Optional column, if arg \code{merge_iei}
is set to \code{TRUE}.
}
}
\description{
Exported files from Synaptosoft's Mini Analysis software (MA, see ref) are saved as ASCII text file
(.asc), this function will convert (.asc) files into a \code{data.frame}.
}
\details{
Mini Analysis software (MA) independently analyzes inter-event interval from the events file, this function
will combine the event file (.asc) with the corresponding inter-event interval (IEI) text file (.txt) if arg
\code{merge_iei} is set to TRUE.
}
\note{
Several columns are set to the base type "character" to avoid generating \code{NA} on number values containing
commas and periods, e.g. \code{1,200.00}, on import.
}
\references{
\url{http://www.synaptosoft.com/MiniAnalysis/}
}
