% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asc_to_tibble}
\alias{asc_to_tibble}
\title{Convert .ASC File Into A Tibble Data Frame}
\usage{
asc_to_tibble(file_path, merge_iei = FALSE)
}
\arguments{
\item{file_path}{Character, path to file. User can call \code{file.choose} to interactively select the file.}

\item{merge_iei}{Logical, if TRUE .asc and .txt files with matching names will be merged. Default is \code{FALSE}.}
}
\value{
A tibble with column names that match default columns within Synaptosoft's Mini Analysis software. Column names are
as follows:
\itemize{
\item \code{ma_row}: Integer, corresponds to the row number generated in Mini Analysis. It is not recommended for use
as a reference/key value.
\item \code{rec_time_ms}: Double, continuous time value that is determined from the recording digitization value.
\item \code{amplitude} Double, amplitude of the detected event.
\item \code{rise_ms} Double, rise time in milliseconds.
\item \code{decay_ms} Double, decay time in milliseconds.
\item \code{area} Double
\item \code{baseline} Double
\item \code{noise} Double
\item \code{group}: Integer (0 to 255), user-assigned group number within MA.
\item \code{channel}: Integer, corresponds to the recording channel as defined by the user's DAC and digitizer.
\item \code{x10_90_rise} Double
\item \code{halfwidth} Double
\item \code{rise50} Double
\item \code{peak_direction} Integer
\item \code{burst} Integer
\item \code{burste} Integer
\item \code{x10_90slope} Numeric
\item \code{rel_time} Double
\item \code{iei}: Double, inter-event interval in milliseconds calculated in MA. Optional column, if arg \code{merge_iei}
is set to \code{TRUE}.
}
}
\description{
Exported files from Synaptosoft's Mini Analysis software (MA, see ref) are saved as ASCII text file
(.asc), this function will convert (.asc) files into a \code{tibble}.
}
\details{
Mini Analysis software (MA) independently analyzes inter-event interval from the events file, this function
will combine the event file (.asc) with the corresponding inter-event interval (IEI) text file (.txt) if arg
\code{merge_iei} is set to TRUE.
}
\references{
\url{http://www.synaptosoft.com/MiniAnalysis/}
}
