% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated_utils.R
\name{batch_convert_asc_to_csv}
\alias{batch_convert_asc_to_csv}
\title{Convert .ASC Files To .cvs Files}
\usage{
batch_convert_asc_to_csv(file_list, merge_iei = FALSE, path = NULL)
}
\arguments{
\item{file_list}{Character, vector of files to process, e.g. \code{matools_env$files_in_folder}.}

\item{merge_iei}{Logical, if TRUE asc and txt files with matching names will be merged.
Default is \code{FALSE}.}

\item{path}{Character, with the location of data directory containing .asc files. Default is NULL
and matools_env$directory_data will be used.}
}
\value{
Exports *.csv file to directory specified in \code{path}, column
names match default columns within Synaptosoft's Mini Analysis software.
}
\description{
Exported files from Synaptosoft's Mini Analysis software (MA,
see ref) are saved as ASCII text file (.asc), this function provides an
user-friendly .csv file for data analysis and viewing. CSV files are
automatically generated in a '\code{WORKING_DIR/data_csv/}' folder.
}
\seealso{
\link{set_data_directory} to auto generate \code{directory_data} in
\code{matools_env}
}
