% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{calculate_stimulus_times}
\alias{calculate_stimulus_times}
\title{Calculate Stimulus Times}
\usage{
calculate_stimulus_times(stimulus_time_first, stimulus_count, stimulus_isi)
}
\arguments{
\item{stimulus_time_first}{Numeric, time of the first stimulus in ms}

\item{stimulus_count}{Numeric, number of stimuli}

\item{stimulus_isi}{Numeric, interstimulus interval in ms, e.g. 25 Hz = 40 ms}
}
\value{
returns a vector of stimulus times to the \code{matools_env}:
\itemize{
\item \code{time_of_stimuli} vector with stimulus times
}
}
\description{
The output can be used to determine the # of stimuli and
timing for subsequent functions.
}
\examples{
\dontrun{
library(matools)

calculate_stimulus_times(
  matools_env$stimulus_time_of_first,
  matools_env$stimulus_count,
  matools_env$stimulus_isi
)
}

}
