% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{condition_sweep_lut}
\alias{condition_sweep_lut}
\title{Generate Condition Sweep Lookup Table (LUT)}
\usage{
condition_sweep_lut(df, condition_names, column_ref = "sweep")
}
\arguments{
\item{df}{Data Frame or Tibble, with the following columns:
* sweep OR another reference column, e.g. trial
* condition}

\item{condition_names}{Character vector with condition names.}

\item{column_ref}{Character, reference column to generate LUT. Defaults to \code{"sweep"}.}
}
\value{
Tibble, with the following columns:
* \code{condition} Character.
* \code{condition_start} Integer, first sweep for a condition.
* \code{condition_end} Integer, last sweep for a condition.
* \code{delta} Integer, number of sweeps in a condition.
}
\description{
Generate Condition Sweep Lookup Table (LUT)
}
\details{
This function will determine the first and last sweep numbers for all conditions, returning a
LUT with the columns, \code{condition_start} and \code{condition_end}. It also calculates the
number of sweeps for each condition in the column, \code{delta}.
}
