% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated_utils.R
\name{create_project_folders}
\alias{create_project_folders}
\title{Creates plot folders within the parent directory}
\usage{
create_project_folders(experiments)
}
\arguments{
\item{experiments}{Character vector with types of experiments, i.e.
experiment_types.}
}
\value{
System folders with the following names:
\itemize{
\item \code{boxplots}
\item \code{scatterplots}
\item \code{histograms}
\item \code{rasterplots}
\item \code{heatplots}
}
}
\description{
Automatic plotting functions save outputs, e.g. *.pdf, into these folders.
}
\note{
Deprecated, use tibble structure.
}
