% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated_utils.R
\name{create_summary_array}
\alias{create_summary_array}
\title{Creates empty experiment summary arrays}
\usage{
create_summary_array(parameters, y, z)
}
\arguments{
\item{parameters}{The output from \code{\link{set_experiment_parameters}} Experiment parameters
data frame.}

\item{y}{Character, user defined name(s) of output (i.e. "analysis_summary")}

\item{z}{Character, vector with types of experiments (i.e. matools_env$experiment_types).}
}
\value{
Nested empty lists segregated by experiment type with corresponding filenames. Created
in the global environment.
}
\description{
Used to store outputs from individual data files (.asc) as they
are processed, allowing for grouped analysis and/or plotting functions at a
later time. Used in conjunction with \link{fill_summary_array}.
}
\examples{
\dontrun{
library(matools)

summary_array_names <- c("analysis_summary", "spike_counts_all_cells")

for (name in summary_array_names) {
  create_summary_array(
    matools_env$parameters,
    name,
    matools_env$experiment_types
  )
}


rm(list = summary_array_names)
}

}
\seealso{
\link{fill_summary_array}
}
