% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{edgefix}
\alias{edgefix}
\title{edgefix}
\usage{
edgefix(gg)
}
\arguments{
\item{gg}{gGraph object to paste edge CN's onto}
}
\value{
gGraph object with pasted edge CN's
}
\description{
fills in missing edge cn (NA after disjoin), then loosefixes output before returning
}
\author{
andrew ma
}
