% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated_utils.R
\name{fill_summary_array}
\alias{fill_summary_array}
\title{Fill summary array}
\usage{
fill_summary_array(id, filename, event_summary, array_blank, array_name)
}
\arguments{
\item{id}{Character, with Experiment id}

\item{filename}{Character, with single filename}

\item{event_summary}{Data Frame}

\item{array_blank}{Object, summary array to be filled. Generated using
\code{create_summary_array}.}

\item{array_name}{String, name of output, e.g "analysis_summary"}
}
\value{
Appended nested lists segregated by experiment type with corresponding filenames.
Revised in the global environment.
}
\description{
Used to store modified dataframes and/or analysis summaries in a
nested list structure - allows the user to generate plots based on
individual or groups of experiments. This is function is soft depreciated
}
\examples{
\dontrun{
library(matools)

summary_array_names <- c("analysis_summary", "spike_counts_all_cells")

for (name in summary_array_names) {
  create_summary_array(
    matools_env$parameters,
    name,
    matools_env$experiment_types
  )
}

fill_summary_array(
  matools_env$experiment_id,
  "file_name",
  df,
  analysis_summary,
  "analysis_summary"
)


rm(list = summary_array_names)
}

}
\seealso{
\link{create_summary_array}
}
