% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{gap_collapse}
\alias{gap_collapse}
\title{gap_collapse}
\usage{
gap_collapse(bp, gap = 50, min.support = 3L, ignore.strand = TRUE)
}
\arguments{
\item{bp}{GRanges of width-1 breakpoints}

\item{gap}{(default = 50) integer for largest gap size to collapse}

\item{min.support}{(default = 3) integer for minimum number of breakpoints to support a cluster}

\item{ignore.strand}{(default = TRUE) logical for whether to ignore strand when collapsing}
}
\value{
GRanges of collapsed breakpoints with metadata col \code{support} indicating number of breakpoints supporting each collapsed breakpoint
}
\description{
gap-collapse a set of width-1 breakpoints into representative cut positions
}
