% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{get.freeze}
\alias{get.freeze}
\title{get.freeze}
\usage{
get.freeze(gg)
}
\arguments{
\item{gg}{gGraph object to get frozen nodes from}
}
\value{
vector of node id's to be frozen
}
\description{
Gets nodes with only REF edges to be frozen in sample.gwalks; returns a vector of the node id's
}
\author{
andrew ma
}
