% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{get_readL}
\alias{get_readL}
\title{get_readL}
\usage{
get_readL(reads)
}
\arguments{
\item{reads}{GRanges or GRangesList of reads}
}
\value{
list with read length distribution and mean read length
}
\description{
outputs read length distribution and mean read length from reads (just uses $qwidth param)
}
\author{
andrew ma
}
