% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{ggplot_color_hue}
\alias{ggplot_color_hue}
\title{Generate Color Hue List}
\usage{
ggplot_color_hue(color_count)
}
\arguments{
\item{color_count}{Numeric, number of colors to generate}
}
\value{
Character vector with six digit hex color value
}
\description{
Manually generating color values allows for consistent
assignment to individual experiment conditions across all df/experiments -
independent of events occurring within a given condition. ggplot's auto
color can assign different colors across df/experiments if a individual df
is missing events.
}
