% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{gr.breaks.ordered}
\alias{gr.breaks.ordered}
\title{gr.breaks.ordered}
\usage{
gr.breaks.ordered(bps = NULL, query = NULL)
}
\arguments{
\item{bps}{GRanges of width 1, locations of the breakpoints; if any element width
larger than 1, both boundary will be considered individual breakpoints}

\item{query}{a disjoint GRanges object to be broken}
}
\value{
Granges disjoint object at least the same length as query,
with metadata col \code{qid} indicating input index where new segment is from and
\code{node_ord} indicating order of new resultant segments
}
\description{
Break GRanges at given breakpoints into disjoint gr
edit: returns GRanges in original order
}
\author{
Xiaotong Yao, andrew ma
}
