% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_experiment_parameters}
\alias{import_experiment_parameters}
\title{Importing Experiment Parameters}
\usage{
import_experiment_parameters(file = NULL)
}
\arguments{
\item{file}{Data.Frame or Character (PATH to .csv file). Users can pass a data frame or tibble into this function or
the full PATH to a .csv file. This allows for the use of a remote/cloud-based file, e.g. \code{googlesheets}, or
local file. Default is \code{NULL} and will prompt the user to interactively select the file.}
}
\value{
Environment objects (variables) to \code{matools_env}:
\itemize{
\item \code{parameters}
\item \code{experiment_types}
}
}
\description{
Importing experiment parameters from a static file allows for batch processing of experiments that may
have asymmetries, e.g. drug, stimulus, experiment id. User should manually fill in a "experiment_parameters.csv"
file. The function, \link{create_parameter_template}, can be called to generate a template file.
}
\examples{
\dontrun{
library(matools)

### Passing the PATH to a local file ###
local_file_path <- "/path/to/my/parameters.csv"
import_experiment_parameters(file = local_file_path)

### Using a publicly shared Google Sheet ###
googlesheets4::gs4_deauth()
public_sheet <- "https://docs.google.com/spreadsheets/d/YOUR_SHEET_HASH_HERE/edit?usp=sharing"

# Convert the sheet to a data frame
parameters <-
  googlesheets4::read_sheet(public_sheet)
import_experiment_parameters(file = parameters)
}

}
\seealso{
\link{create_parameter_template} to generate a csv template.
}
