% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{insert_rows_for_missing_sweeps}
\alias{insert_rows_for_missing_sweeps}
\title{Add Blank Rows To Data Frame For Missing Sweeps/Trials}
\usage{
insert_rows_for_missing_sweeps(df, sweep_count)
}
\arguments{
\item{df}{Data Frame or Tibble, with the following columns:
\itemize{
\item sweep
}}

\item{sweep_count}{Numeric, total number of sweeps.}
}
\value{
Appended data.frame with rows for sweeps/trials without events, only the ordered factor column, \code{sweep},
will contain a integer values, all other row values are \code{NA}.
}
\description{
Events may not occur on every sweep/trial, this will add a blank row for sweeps without events.
Thus, creating continuous sweep data allowing for addition of condition tags and/or analysis by sweep number.
}
\examples{
simple_df <- data.frame(
  rec_time_ms = c(400, 4400),
  sweep = ordered(c(1, 5), levels = c(1:5))
)

insert_rows_for_missing_sweeps(
  df = simple_df,
  sweep_count = 5
)

}
\seealso{
\itemize{
\item \link{add_sweep_number_to_rows}
}
}
