% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{jac}
\alias{jac}
\title{jac}
\usage{
jac(a, b)
}
\arguments{
\item{a}{(vector) first set}

\item{b}{(vector) second set}
}
\value{
Jaccard index
}
\description{
computes the Jaccard index between two sets
}
