% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{jct_filt}
\alias{jct_filt}
\title{jct_filt}
\usage{
jct_filt(
  bp,
  anchors,
  gap = 50,
  min.support = 3L,
  ignore.strand = TRUE,
  keep.unsupported.anchors = TRUE
)
}
\arguments{
\item{bp}{GRanges of width-1 breakpoints}

\item{anchors}{GRanges of width-1 breakpoints to snap onto first}

\item{gap}{(default = 50) integer for largest gap size to collapse}

\item{min.support}{(default = 3) integer for minimum number of breakpoints to support a cluster}

\item{ignore.strand}{(default = TRUE) logical for whether to ignore strand when collapsing}

\item{keep.unsupported.anchors}{(default = TRUE) logical for whether to keep anchors with no snapped breakpoints}
}
\value{
GRanges of collapsed breakpoints with support mcol
}
\description{
snap read breakpoints onto authoritative anchors first, then gap-cluster the rest
}
