% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadIndexDataQDW.R
\name{loadIndexDataQDW}
\alias{loadIndexDataQDW}
\title{Load data from QDW}
\usage{
loadIndexDataQDW(envir, compKey)
}
\arguments{
\item{envir}{choose QDW environment. Defaults to prod}

\item{compKey}{vector of composite keys. e.g. c('DDL|PI|GSALSPT','DDL|PI|CGSYSPT')}
}
\description{
Load data from the Schroders Quant Data Warehouse (QDW)
}
\examples{
compKey <- c('DDL|PI|GSALSPT','DDL|PI|CGSYSPT')
out <- loadIndexDataQDW(envir = 'prod', compKey)
}
\keyword{QDW}
