% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.adjustFreq.R
\name{ma.adjustFreq}
\alias{ma.adjustFreq}
\title{adjust xts frequency - assumes a less frequent periodicity is required}
\usage{
ma.adjustFreq(z, freq, day)
}
\arguments{
\item{x}{}
}
\description{
adjust xts frequency - assumes a less frequent periodicity is required
}
