% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.calculateReturns.R
\name{ma.calculateReturns}
\alias{ma.calculateReturns}
\title{Calculate perodic returns}
\usage{
ma.calculateReturns(x, nPer = 1)
}
\arguments{
\item{x}{xts object of prices}

\item{nPer}{choose Nper period returns}
}
\description{
Calculate perodic returns from xts object of prices
}
\examples{

}
\keyword{performance}
