% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.size2CommonStartDate.R
\name{ma.size2CommonStartDate}
\alias{ma.size2CommonStartDate}
\title{size xts to first commom start date}
\usage{
ma.size2CommonStartDate(z)
}
\arguments{
\item{x}{}
}
\description{
size xts to first commom start date
}
