% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{map.fine}
\alias{map.fine}
\title{map.fine}
\usage{
map.fine(gws, gg, return.gw = FALSE)
}
\arguments{
\item{gws}{gWalk object to be mapped}

\item{gg}{gGraph object to map to}

\item{return.gw}{logical, if TRUE returns a new gWalk object, if FALSE returns a list of mapped snode.id's}

\item{minsize}{minimum size of a node to be included in the output}
}
\value{
list of mapped snode.id's or a new gWalk object
}
\description{
maps gWalks to a finer reference gGraph's node.id's (e.g. read-disjoined ref graph)
and returns a new gWalk object with the finer node.id's
}
\author{
andrew ma
}
