\name{matools-package}
\alias{matools-package}
\alias{matools}

\docType{package}
\title{Motion analysis tools}
\description{
\strong{matools} provides functions for motion analysis in the R language.
}
\details{
\tabular{ll}{
Package: \tab matools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-03-17\cr
License: \tab GPL-2\cr
}
}
\author{
Aaron Olsen

Maintainer: Aaron Olsen \email{aarolsen@gmail.com}
}
\keyword{ package }