% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{pad.walk}
\alias{pad.walk}
\title{pad.walk}
\usage{
pad.walk(gw, gap.thresh = 2)
}
\arguments{
\item{gap.thresh}{(default = 2) threshold for gap sizes to be filled}
}
\value{
gWalk object
}
\description{
fills in trivial gaps between walk nodes; walks along path and can pad asymmetrically
}
\author{
andrew ma
}
