% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_event_avg.R
\name{plot_event_avg}
\alias{plot_event_avg}
\title{Plot Event Value (Average)}
\usage{
plot_event_avg(
  df,
  x_axis,
  y_axis,
  plot_grouping,
  plot_mean = FALSE,
  weighted_mean = FALSE,
  sd_value = NULL,
  experiment_name = NULL,
  x_label = NULL,
  y_label = NULL,
  ymax = NA,
  ...
)
}
\arguments{
\item{df}{Data frame or Tibble, grouped experiment data set.}

\item{x_axis}{Character - unquoted, non-standard evaluation. Column name to use for x-axis.}

\item{y_axis}{Character - unquoted, non-standard evaluation. Column name to use for y-axis.}

\item{plot_grouping}{Character, column name to use for grouping.}

\item{plot_mean}{Boolean, add group mean to plot. Default is \code{FALSE}.}

\item{weighted_mean}{Boolean, plot weighted mean, requires df column labelled \code{'w'} of weights \eqn{w}.
Default is \code{FALSE}.}

\item{sd_value}{Optional character, std deviation measurement to use, accepts "sd" or "sem". Default is \code{NULL}.}

\item{experiment_name}{Optional character, add a subtitle with experiment name. Default is \code{NULL}.}

\item{x_label}{Optional character, custom label for x-axis. Default \code{NULL}, will use the x_axis column name.}

\item{y_label}{Optional character, custom label for y-axis. Default \code{NULL}, will use the y_axis column name.}

\item{ymax}{Optional integer, set a y-axis limit. Default is \code{NULL} and the y-axis will be automatically
determined by ggplot.}

\item{...}{Optional vargs for additional ggplot function calls, e.g. \code{ylim()}, \code{ylab()}.}
}
\value{
A ggplot2 object with user defined features.
}
\description{
A highly abstracted plot that can be used to plot an average value within and across experiments. This
is useful for examining features of synaptic events, e.g. amplitude, failure rate, potency, paired-pulse ratio.
The plot requires pre-processing of the feature(s) of interest and a column with unique values to group by, i.e.
experiment recording IDs.
}
\details{
The weight column, \code{w}, is assumed to frequency weights - where a weight equals the number of occurrences.
Weighted mean and sd are calculated using the following formulas:
\itemize{
\item weighted mean:
\eqn{\bar{x}=\frac{\sum\limits_{i=1}^n w_i x_i}{\sum\limits_{i=1}^n w_i}}
\item weighted sd: \eqn{\sqrt{\hat{\sigma}^2_\mathrm{w}}}
\item unbiased variance:
\eqn{\hat{\sigma}^2_\mathrm{w}=\frac{\sum\limits_{i=1}^N w_i \left(x_i - \mu^{*}\right)^2 }{\sum_{i=1}^N w_i - 1}}
}
}
\examples{
simple_tbl <-
  data.frame(
    recording_id = rep(1:2, each = 3),
    experiment_id = rep("example", 6),
    condition = rep(c("control", "drug_1", "drug_2"), 2),
    amplitude = c(1, 0.5, 1, 2, 1, 2)
  )

plot_event_avg(
  df = simple_tbl,
  x_axis = condition,
  y_axis = amplitude,
  plot_grouping = "recording_id",
  plot_mean = TRUE,
  sd_value = "sem",
  experiment_name = "example",
  ymax = 3
) + ggplot2::coord_fixed(ratio = 3)

# remove average and add custom labels
plot_event_avg(
  df = simple_tbl,
  x_axis = condition,
  y_axis = amplitude,
  plot_grouping = "recording_id",
  plot_mean = FALSE,
  experiment_name = "example",
  x_label = "custom x-label",
  y_label = "custom y-label",
  ymax = 3
)

# plot a weighted average
weighted_tbl <-
  data.frame(
    recording_id = rep(1:2, each = 3),
    experiment_id = rep("example", 6),
    condition = rep(c("control", "drug_1", "drug_2"), 2),
    amplitude = c(1, 0.5, 1, 2, 1, 2),
    w = rep(c(1, 100), each = 3)
  )

plot_event_avg(
  df = weighted_tbl,
  x_axis = condition,
  y_axis = amplitude,
  plot_grouping = "recording_id",
  plot_mean = TRUE,
  weighted_mean = TRUE,
  sd_value = "sd",
  experiment_name = "plot weighted mean",
  x_label = "condition",
  y_label = "amplitude",
  ymax = 3,
)

}
\references{
\itemize{
\item \href{https://en.wikipedia.org/wiki/weighted_arithmetic_mean}{Wikipedia: Weighted arithmetic mean}
\item \href{https://github.com/harrelfe/Hmisc/issues/22#issuecomment-94307602}{GitHub: harrelfe/Hmisc - issues #22}
}
}
\seealso{
\itemize{
\item \code{\link[stats:weighted.mean]{stats::weighted.mean()}}
\item \code{\link[Hmisc:wtd.stats]{Hmisc::wtd.var()}}
}
}
