% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heat Map of Event Amplitudes}
\usage{
plot_heatmap(df, filename, condition_names, max_amplitude = NULL)
}
\arguments{
\item{df}{Data Frame or Tibble with the following columns:
\itemize{
\item sweep
\item amplitude
\item condition
\item stimulus
\item event index
}}

\item{filename}{Character, name of the file.}

\item{condition_names}{Character vector with condition names.}

\item{max_amplitude}{Integer, maximum value to use for the heat map. If set less than the maximum
amplitude in the data frame, the value will be changed and a warning is returned. Defaults to
\code{NULL} and will automatically set the value to the df's max value.}
}
\value{
ggplot object
}
\description{
Plot Heat Map of Event Amplitudes
}
\details{
A blank white space is used for missing events (amplitude == NA).
}
\examples{
simple_df <-
  data.frame(
    sweep = rep(1:20, each = 2),
    amplitude = c(rnorm(20, 100, 20), rnorm(20, 50, 10)),
    condition = rep(c("control", "drug1"), rep(20, 2)),
    stimulus = rep(c(1:2), 20),
    event_index = rep(1, 40)
  )

plot_heatmap(
  df = simple_df,
  filename = "example",
  condition_names = c("control", "drug1")
)

}
\seealso{
\itemize{
\item \link{align_condition_sweeps}
}
}
