% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_probability.R
\name{plot_max_probability}
\alias{plot_max_probability}
\title{Plot Max Probability Across Experiments}
\usage{
plot_max_probability(
  df,
  x_axis_condition,
  y_axis_condition,
  metric = "probability",
  xmax = 1,
  ymax = 1,
  x_label = NULL,
  y_label = NULL
)
}
\arguments{
\item{df}{Data Frame or Tibble with the following columns:
\itemize{
\item cell_id
\item experiment_id
\item condition
\item P() metric from \code{create_histogram_df()}, e.g. density, probability.
}}

\item{x_axis_condition}{Character, condition value to use on x-axis, e.g. "control".}

\item{y_axis_condition}{Character, condition value to use on y-axis, e.g. "drug".}

\item{metric}{Character, P() metric to plot, e.g. density, probability. Defaults to
"probability".}

\item{xmax}{Optional integer, maximum x-value to plot, values are rounded the nearest 1. Default
is \code{1}, ymin is fixed at \code{0}.}

\item{ymax}{Optional integer, maximum y-value to plot, values are rounded the nearest 1. Default
is \code{1}, ymin is fixed at \code{0}.}

\item{x_label}{Optional character, custom label for the x-axis. Default \code{NULL}, will use the
column name from \code{x_axis_condition}.}

\item{y_label}{Optional character, custom label for the y-axis. Default \code{NULL}, will use the
column name from \code{y_axis_condition}.}
}
\value{
ggplot2 object
}
\description{
This will plot the maximum value specified by \code{metric} across two conditions.
}
\note{
This function is intended to use with aggregated data in a long format that is extracted
from \code{create_histogram_df}.
}
\examples{
library(magrittr, include.only = "\%>\%")

bins <- seq(0.5, 39.5, by = 1)
n_bins <- length(bins)
conditions <- c("control", "drug")
n_conditions <- length(conditions)

# example output from create_histogram_df()
simple_tbl <-
  tibble::tibble(
    cell_id = "cell",
    experiment_id = "example",
    condition = rep(conditions, each = n_bins),
    h_bin_centers = rep(bins, n_conditions),
    probability = 0,
    density = 1
  ) \%>\%
  dplyr::mutate(probability = dplyr::case_when(
    condition == "control" & h_bin_centers \%in\% c(1.5, 11.5) ~ 0.75,
    condition == "drug" & h_bin_centers \%in\% c(1.5, 11.5) ~ 0.5,
    .default = 0
  ))

plot_max_probability(
  simple_tbl,
  x_axis_condition = "control",
  y_axis_condition = "drug",
  metric = "probability",
  xmax = 1,
  ymax = 1
)

}
\seealso{
\itemize{
\item \code{\link[=create_histogram_df]{create_histogram_df()}}
\item \code{\link[=plot_probability]{plot_probability()}}
\item \code{\link[=analysis_evoked_ap]{analysis_evoked_ap()}} for a workflow example that aggregates create_histogram_df.
}
}
