% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spike_raster.R
\name{plot_spike_raster}
\alias{plot_spike_raster}
\title{Generate Spike Raster Plot}
\usage{
plot_spike_raster(
  df,
  filename,
  id,
  time_of_stim,
  condition_names,
  auto_window = TRUE,
  auto_window_size = 10,
  auto_window_symmetric = FALSE,
  manual_condition_color = NULL,
  plot_conditions_only = TRUE
)
}
\arguments{
\item{df}{Data Frame or Tibble with the following columns:
\itemize{
\item sweep
\item time_ms
\item condition
}}

\item{filename}{Character, a single file name.}

\item{id}{Character, experiment id.}

\item{time_of_stim}{List, stimulus times in milliseconds.}

\item{condition_names}{Character vector with condition titles.}

\item{auto_window}{Boolean, automatically set the x-axis values based upon the first and last
stimuli. Default is \code{TRUE} and will use the args \code{auto_window_size} and
\code{auto_window_symmetric} to set the x-axis. If \code{FALSE}, the entire x-axis is plotted
using the values c(0, \code{NA}).}

\item{auto_window_size}{Integer, the amount of time in milliseconds to plot after the last
stimulus. Default is \code{10} ms.}

\item{auto_window_symmetric}{Boolean, will set the x-axis to plot \code{auto_window_size} before
and after the first and last stimuli. Default is \code{FALSE}, and will start the x-axis at the
first stimuli.}

\item{manual_condition_color}{Character, vector of color values. Default is \code{NULL}, which
will force arg \code{auto_condition_color} to \code{TRUE}, automatically generating color
values.}

\item{plot_conditions_only}{Boolean, Default is \code{TRUE}.}
}
\value{
ggplot object
}
\description{
Generate Spike Raster Plot
}
\examples{
conditions <- c("control", "drug_1", "drug_2")
simple_df <-
  data.frame(
    sweep = rep(1:120, each = 3),
    time_ms = rep(c(280, 290, 300), times = 120),
    condition = rep(conditions, each = 120)
  )

plot_spike_raster(
  df = simple_df,
  filename = "example",
  id = "experiment_id",
  time_of_stim = c(278.18, 288.18, 298.18),
  condition_names = conditions,
  auto_window = TRUE,
  auto_window_size = 30,
  auto_window_symmetric = FALSE,
  plot_conditions_only = TRUE
)

}
\seealso{
\itemize{
\item \link{standardize_event_time} to auto create \code{time_ms} column.
\item \link{ggplot_color_hue} to see how 'auto_condition_color' is determined.
\item \link{condition_sweep_lut}
}
}
