% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{read2node}
\alias{read2node}
\title{read2node}
\usage{
read2node(gr = NULL, gg = NULL, gap = 2, inv = 50)
}
\arguments{
\item{gr}{GRanges object of a single read from BAM}

\item{gg}{gGraph object of reference graph that you want to map node.id from}

\item{gap}{(default = 2) integer for largest gap size to close}

\item{inv}{(defaul = 50) integer for largest inversion interval to toss}
}
\value{
gWalk object
}
\description{
maps a read GRanges to a sequence of node id's built from the reference graph
}
\author{
andrew ma
}
