% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{ref_from_gg}
\alias{ref_from_gg}
\title{ref_from_gg}
\usage{
ref_from_gg(gg, cn = 2)
}
\arguments{
\item{gg}{gGraph object}

\item{cn}{(default = 2) integer for the copy number of the reference karyotype}
}
\value{
gg gGraph that is diploid and unrearranged, with the same nodes
}
\description{
Generate a reference karyotype from an event graph
}
