% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_utils.R
\name{reset_pkg_environment}
\alias{reset_pkg_environment}
\title{Reset 'matools' Environment}
\usage{
reset_pkg_environment()
}
\value{
A new \code{\link[=environment]{environment()}}, \code{matools_env}, in the \code{.GlobalEnv} namespace.
}
\description{
This is a wrapper for \code{set_pkg_environment}
}
\examples{
library(matools)

# create an environment and add objects ----
set_pkg_environment(force_new = TRUE)
matools_env$parameter_1 <- TRUE
matools_env$parameter_2 <- FALSE
ls(matools_env)

# reset the environment ----
reset_pkg_environment()
rm(matools_env, envir = .GlobalEnv)

}
\seealso{
\code{\link[=set_pkg_environment]{set_pkg_environment()}}
}
