% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{rminv.walk}
\alias{rminv.walk}
\title{rminv.walk}
\usage{
rminv.walk(gw, inv.thresh = 1)
}
\arguments{
\item{gw}{input gWalk object}

\item{inv.thresh}{(default = 1) threshold for inversions to be dropped}
}
\value{
gWalk object
}
\description{
remove trivial inversions under a size threshold for walks
}
\author{
andrew ma
}
