% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{sequential_condition_sweeps}
\alias{sequential_condition_sweeps}
\title{Sequential Condition Sweeps}
\usage{
sequential_condition_sweeps(df, condition_names)
}
\arguments{
\item{df}{Data Frame or Tibble, with the following columns:
* sweep
* condition}

\item{condition_names}{Character vector with condition names.}
}
\value{
Data Frame or Tibble, with the additional column:
* trial
}
\description{
Sequential Condition Sweeps
}
\details{
Creates a 'trial' column with sequential trial numbers across all conditions. The original
sweep value is not overwritten.
}
\seealso{
* \link{align_condition_sweeps}
* \link{plot_spike_raster}
}
