% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_data_directory}
\alias{set_data_directory}
\title{Set Data Directory}
\usage{
set_data_directory(file_path = NULL)
}
\arguments{
\item{file_path}{Optional path to data directory. Defaults to \code{NULL} and opens GUI for directory selection.}
}
\value{
Environment variables to \code{matools_env}:
\itemize{
\item \code{directory_data}
\item \code{directory_parent}
\item \code{files_in_folder}
}
}
\description{
Extracts data directory, it's parent and names of data files (*.asc) within the directory. This will
aid with creating an analysis pipeline.
}
\examples{
library(matools)

set_pkg_environment()
set_data_directory(file_path = getwd())
print(matools_env$directory_data)

}
\seealso{
\link{set_pkg_environment} to generate the required matools_env
}
