% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_experiment_parameters}
\alias{set_experiment_parameters}
\title{Set Experiment Parameters}
\usage{
set_experiment_parameters(
  parameters,
  filename,
  condition_names = lifecycle::deprecated()
)
}
\arguments{
\item{parameters}{Experiment parameters df imported using
\code{import_experiment_parameters} function.}

\item{filename}{Character, single file name}

\item{condition_names}{\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}:Character
vector with condition titles}
}
\value{
environment variables to \code{matools_env}:
\itemize{
\item \code{experiment_id} chr
\item \code{stimulus_time_of_first} dbl
\item \code{stimulus_count} int
\item \code{stimulus_isi} int
\item \code{stimulus_time_of_recovery} dbl
\item \code{event_threshold} int
\item \code{sweep_total} int
\item \code{sweep_duration_sec} int
\item \code{condition_names} chr
\item \code{condition_starts} list
\item \code{condition_ends} list
\item \code{condition_drug_starts}{lifecycle::badge("deprecated")}
}
}
\description{
Pulls file specific parameters (i.e. drug start times) from the parameters data frame, allowing for
analysis f() based upon individual experiment settings.
}
