% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_utils.R
\name{set_pkg_environment}
\alias{set_pkg_environment}
\title{Create An New 'matools' Environment To Store Analysis Parameters}
\usage{
set_pkg_environment(force_new = FALSE)
}
\arguments{
\item{force_new}{Boolean, when TRUE a new matools_env will be created,
overwriting all existing parameters. Default is FALSE.}
}
\value{
A new \code{\link[=environment]{environment()}}, \code{matools_env}, in the \code{.GlobalEnv} namespace.
}
\description{
Provides a location to store package objects and variables,
allowing for easy removal of data analysis parameters within an analysis
pipeline. Note: sets parent environment to \code{env: empty} to avoid scope
escape.
}
\examples{
library(matools)

set_pkg_environment(force_new = TRUE)
print(matools_env)
rm(matools_env, envir = .GlobalEnv)

}
\seealso{
\code{\link[=reset_pkg_environment]{reset_pkg_environment()}}
}
