% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{smooth.cigar}
\alias{smooth.cigar}
\title{smooth.cigar}
\usage{
smooth.cigar(alignments = NULL, smooth.thresh = 50)
}
\arguments{
\item{alignments}{(default = NULL) GRanges or GRangesList of pooled reads}

\item{smooth.thresh}{(default = 50) integer for largest deletion size to smooth}
}
\value{
GRanges or GRangesList of pooled reads with smoothed cigar strings
}
\description{
makes a smoothed cigar by removing I's and converting D's to M's, all w/in a certain width threshold
}
\author{
andrew ma
}
