% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{standardize_event_time}
\alias{standardize_event_time}
\title{Standardize Event Time To Discrete Time}
\usage{
standardize_event_time(df, sweep_duration, time_ref = "rec_time_ms")
}
\arguments{
\item{df}{Data Frame or Tibble with a \code{sweep} column}

\item{sweep_duration}{Numeric, sweep/trial length in seconds.}

\item{time_ref}{Character, event time reference point column name. Defaults to "rec_time_ms".}
}
\value{
Data Frame or Tibble with transformed time column, \code{time_ms}.
}
\description{
For converting continuous time into discrete time series for functions that plot by sweeps or calculate
events by stimuli within sweeps.
}
\examples{
simple_df <- data.frame(
  rec_time_ms = c(400, 4400),
  sweep = ordered(c(1, 5), levels = c(1:5))
)

standardize_event_time(
  df = simple_df,
  sweep_duration = 1,
  time_ref = "rec_time_ms"
)
\dontrun{
df <- standardize_event_time(
  df,
  matools_env$sweep_duration_sec,
  "rec_time_ms"
)
}

}
\seealso{
\itemize{
\item \link{add_sweep_number_to_rows}
}
}
