% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotyping_utils.R
\name{alignment2gw}
\alias{alignment2gw}
\title{alignment2gw}
\usage{
alignment2gw(alignment, verbose = TRUE)
}
\value{
gWalk object with nodes and edges lifted to reference coordinates
}
\description{
chunk of alignments2gg, run this on a single contig and get the gwalk in ref coords
}
\author{
Marcin Imielinski, Joe DeRose, Xiaotong Yao, andrew ma
}
